rename endpoint view file, re #993
