+++ afs/views/analysis_area_and_sample_taking.py
